movement and launch rework
Launch rework:
In order to launch, mass of rocket must be less than the thrust of al engines combined. 

Movement rework:
2 movement modes:
RCS, 
Engine.
to activate engine, press space. This moves the rocket forward. 
in RCS mode, wasd controls to move. if engine and rcs is both on, a and d works, d doesn't (backward movement only with off engine)
RCS must be on for rotation of craft